backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
drf-flex-fields==0.7.5
mysqlclient==1.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
